compression_data_collection/requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
coloredlogs==15.0
cycler==0.10.0
decorator==5.0.9
dill==0.3.3
getch==1.0
gym==0.18.3
humanfriendly==9.1
idna==2.10
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
lxml==4.6.3
lz4==3.1.3
matplotlib==3.4.1
matplotlib-inline==0.1.2
minerl==0.3.7
numpy==1.20.2
opencv-python==4.5.2.52
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pyglet==1.5.15
Pygments==2.9.0
pyparsing==2.4.7
Pyro4==4.80
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.6.3
serpent==1.40
six==1.15.0
tqdm==4.61.0
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
